Refactor Session and Server into separate modules
Having both Structs in the same module was becoming unwieldy.

Split them apart and used `pub use` to pull the Server into a public
scope that the main.rs could use.

Also resulted in a cleaner Session interface with better encapsulation.
